pyproject.toml
[tool.poetry]
name = "timeline-api"
version = "0.1.0"
description = ""
authors = ["mhendricks96 <mhendricks96@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
whitenoise = "^6.1.0"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.1.0"
django-friendship = "^1.9.6"
django-rest-friendship = "^0.2.1"
django-environ = "^0.8.1"
django-cors-headers = "^3.12.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
